% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.m.fun}
\alias{ex.m.fun}
\title{Tutorial 5 m() function}
\format{
Function
}
\usage{
ex.m.fun
}
\description{
The m() function supplied to Optimus() for reproducing Tutorial 5. This is
for the small executable example.
}
\keyword{datasets}
